Fix to setting Order use_shipping_address_for_billing in eloquent
This appears to happen before the casts() method and if the value
property on the html element hydrating this is set to anything other
than "true" this was setting this to bool false.

IIRC the default for HTML checkbox inputs is the string "on" if
checked and "off" if not.