[7.x] Fix `use_shipping_address_for_billing` option when using Database Orders (#1213)
* Fix to setting Order use_shipping_address_for_billing in eloquent

This appears to happen before the casts() method and if the value
property on the html element hydrating this is set to anything other
than "true" this was setting this to bool false.

IIRC the default for HTML checkbox inputs is the string "on" if
checked and "off" if not.

* Add failing test

* Keep the `== 'true'` check

* Also, value should be true when value is "on"

---------

Co-authored-by: Duncan McClean <duncan@duncanmcclean.com>