fix style, added test
On my shop I show all prices without VAT because it is a B2B Site and the Tax is added at checkout.
A Product costs 100 Euro (without Tax) and the cart should show 119 Euro total (with 19 % Tax)
Previously the StandardTaxEngine calculates a total of 116 Euro which is wrong.
I added a Test for this Issue, in 5 of the older tests with 'price_includes_tax' => false the expected numbers are not right.