[7.x] Correct tax calculations when "Prices include tax" is false (standard tax engine) (#1234)
* Fix Tax Calculation with $taxRate->includeInPrice true

Tax calculation was not correct when $taxRate->includeInPrice set to true.

* fix style

* fix style, added test

On my shop I show all prices without VAT because it is a B2B Site and the Tax is added at checkout.
A Product costs 100 Euro (without Tax) and the cart should show 119 Euro total (with 19 % Tax)
Previously the StandardTaxEngine calculates a total of 116 Euro which is wrong.
I added a Test for this Issue, in 5 of the older tests with 'price_includes_tax' => false the expected numbers are not right.

* Update test to reflect actual numbers

* Fix shipping tax calculation

* Correct assertions in tests

---------

Co-authored-by: Duncan McClean <duncan@duncanmcclean.com>